# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Haiku",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/haiku": {
            "post": {
                "tags": [
                    "Haiku"
                ],
                "summary": "Create a new haiku",
                "description": "Create a new haiku entry",
                "operationId": "createHaiku",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "is_hidden",
                                    "theme_id",
                                    "user_id",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Example Title"
                                    },
                                    "is_hidden": {
                                        "type": "boolean",
                                        "example": 0
                                    },
                                    "theme_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Line 1\\nLine 2\\nLine 3\\nLine 4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Haiku created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "maxLength": 255,
                                            "example": "Example Title"
                                        },
                                        "is_hidden": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "theme_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Line 1\\nLine 2\\nLine 3\\nLine 4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "field_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Error message"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/haiku": {
            "get": {
                "tags": [
                    "Haiku"
                ],
                "summary": "Get all haiku",
                "operationId": "35e5086a090bac69b8193447387ac480",
                "responses": {
                    "200": {
                        "description": "List of haiku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Example Title"
                                            },
                                            "is_hidden": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "theme_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Line 1\\nLine 2\\nLine 3\\nLine 4"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/haiku/{slug}": {
            "put": {
                "tags": [
                    "Haiku"
                ],
                "summary": "Update a haiku",
                "description": "Update a haiku entry by its slug",
                "operationId": "a5b4e7584ae40009858bf407577af387",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the haiku to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "is_hidden",
                                    "theme_id",
                                    "user_id",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Updated Title"
                                    },
                                    "is_hidden": {
                                        "type": "boolean",
                                        "example": 0
                                    },
                                    "theme_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated Line 1\\nUpdated Line 2\\nUpdated Line 3\\nUpdated Line 4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Haiku updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "maxLength": 255,
                                            "example": "Updated Title"
                                        },
                                        "is_hidden": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "theme_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Updated Line 1\\nUpdated Line 2\\nUpdated Line 3\\nUpdated Line 4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "field_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Error message"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Haiku"
                ],
                "summary": "Delete a haiku",
                "description": "Deletes a haiku by its slug",
                "operationId": "885a3303aa12b3771416bbe59418f339",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the haiku to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Haiku deleted successfully"
                    },
                    "404": {
                        "description": "Haiku not found"
                    }
                }
            }
        }
    }
}